+++ src/net/fexcraft/app/fmt/porters/MTBImporter.java
@@ -2,7 +2,7 @@ package net.fexcraft.app.fmt.porters;
 
 import net.fexcraft.app.fmt.FMTB;
 import net.fexcraft.app.fmt.porters.PorterManager.InternalPorter;
-import net.fexcraft.app.fmt.ui.general.DialogBox;
+import net.fexcraft.app.fmt.ui_old.general.DialogBox;
 import net.fexcraft.app.fmt.utils.TextureManager;
 import net.fexcraft.app.fmt.utils.Settings.Setting;
 import net.fexcraft.app.fmt.wrappers.*;
