+++ src/net/fexcraft/app/fmt/ui_old/Dialog.java
@@ -1,4 +1,4 @@
-package net.fexcraft.app.fmt.ui;
+package net.fexcraft.app.fmt.ui_old;
 
 import java.util.ArrayList;
