+++ src/net/fexcraft/app/fmt/ui_old/FontRenderer.java
@@ -1,4 +1,4 @@
-package net.fexcraft.app.fmt.ui;
+package net.fexcraft.app.fmt.ui_old;
 
 import java.awt.Color;
 import java.awt.Font;
