+++ src/net/fexcraft/app/fmt/ui_old/editor/TextureEditor.java
@@ -1,8 +1,8 @@
-package net.fexcraft.app.fmt.ui.editor;
+package net.fexcraft.app.fmt.ui_old.editor;
 
-import net.fexcraft.app.fmt.ui.Element;
-import net.fexcraft.app.fmt.ui.general.Button;
-import net.fexcraft.app.fmt.ui.general.TextField;
+import net.fexcraft.app.fmt.ui_old.Element;
+import net.fexcraft.app.fmt.ui_old.general.Button;
+import net.fexcraft.app.fmt.ui_old.general.TextField;
 import net.fexcraft.lib.common.math.RGB;
 
 public class TextureEditor extends Editor {
