+++ src/net/fexcraft/app/fmt/ui_old/general/TextField.java
@@ -1,11 +1,11 @@
-package net.fexcraft.app.fmt.ui.general;
+package net.fexcraft.app.fmt.ui_old.general;
 
 import java.util.ArrayList;
 import java.util.Optional;
 
 import net.fexcraft.app.fmt.FMTB;
-import net.fexcraft.app.fmt.ui.Element;
-import net.fexcraft.app.fmt.ui.FontRenderer;
+import net.fexcraft.app.fmt.ui_old.Element;
+import net.fexcraft.app.fmt.ui_old.FontRenderer;
 import net.fexcraft.app.fmt.utils.GGR;
 import net.fexcraft.lib.common.math.RGB;
