+++ src/net/fexcraft/app/fmt/utils/KeyCompound.java
@@ -10,8 +10,8 @@ import org.lwjgl.input.Keyboard;
 import com.google.gson.JsonObject;
 
 import net.fexcraft.app.fmt.FMTB;
-import net.fexcraft.app.fmt.ui.editor.Editor;
-import net.fexcraft.app.fmt.ui.general.DialogBox;
+import net.fexcraft.app.fmt.ui_old.editor.Editor;
+import net.fexcraft.app.fmt.ui_old.general.DialogBox;
 import net.fexcraft.lib.common.json.JsonUtil;
 
 public class KeyCompound {
